# openapi/tourplanner.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TourPlanner API",
    "version": "v1"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/routes/resolve": {
      "post": {
        "tags": [
          "Routes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveRouteResponse"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/reports/summary": {
      "get": {
        "tags": [
          "Reports"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/reports/tour/{tourId}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "tourId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/reports/export/{tourId}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "tourId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/reports/import": {
      "post": {
        "tags": [
          "Reports"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/tour": {
      "post": {
        "tags": [
          "Tour"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "get": {
        "tags": [
          "Tour"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/tour/{id}": {
      "get": {
        "tags": [
          "Tour"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "put": {
        "tags": [
          "Tour"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "delete": {
        "tags": [
          "Tour"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/tour/search/{searchText}": {
      "get": {
        "tags": [
          "Tour"
        ],
        "parameters": [
          {
            "name": "searchText",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/tourlog": {
      "post": {
        "tags": [
          "TourLog"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/tourlog/{id}": {
      "get": {
        "tags": [
          "TourLog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "put": {
        "tags": [
          "TourLog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "delete": {
        "tags": [
          "TourLog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/tourlog/bytour/{tourId}": {
      "get": {
        "tags": [
          "TourLog"
        ],
        "parameters": [
          {
            "name": "tourId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourLogDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourLogDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourLogDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "minLength": 6,
            "type": "string"
          }
        }
      },
      "ResolveRouteRequest": {
        "required": [
          "fromLatitude",
          "fromLongitude",
          "toLatitude",
          "toLongitude",
          "transportType"
        ],
        "type": "object",
        "properties": {
          "fromLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "fromLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "toLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "toLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "transportType": {
            "type": "string"
          }
        }
      },
      "ResolveRouteResponse": {
        "required": [
          "distance",
          "duration"
        ],
        "type": "object",
        "properties": {
          "distance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "duration": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "TourDto": {
        "required": [
          "name",
          "description",
          "from",
          "to",
          "transportType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "imagePath": {
            "type": [
              "null",
              "string"
            ]
          },
          "routeInformation": {
            "type": [
              "null",
              "string"
            ]
          },
          "distance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "estimatedTime": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "transportType": {
            "type": "string"
          },
          "tourLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourLogDto"
            }
          },
          "popularity": {
            "type": [
              "null",
              "string"
            ]
          },
          "isChildFriendly": {
            "type": "boolean"
          },
          "averageRating": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "TourLogDto": {
        "required": [
          "comment",
          "difficulty",
          "totalDistance",
          "totalTime",
          "rating"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tourId": {
            "type": "string",
            "format": "uuid"
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": [
              "null",
              "string"
            ]
          },
          "difficulty": {
            "maximum": 5,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalDistance": {
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalTime": {
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      }
    },
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "name": ".AspNetCore.Identity.Application",
        "in": "cookie"
      }
    }
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Routes"
    },
    {
      "name": "Reports"
    },
    {
      "name": "Tour"
    },
    {
      "name": "TourLog"
    }
  ]
}